This finally populates the canvas on init()
Final solution was window.load()

Lots was changed to bring us these solutions; leading to the inelegant pushing of <imgs to after the <canvas> creating horizontal scrolling.  Either try and go back to the way it was done before but retain the window.load (PREFFERED) or put the <img> in a holding <div>  with horizontal-scrolling-x: none.